Home

ValuedTuples.jl

ValuedTuples.ValuedTuple

Base.merge

ValuedTuples.delete

ValuedTuples.match_index

ValuedTuples.value_names

ValuedTuples.@VT

ValuedTuples.@value

ValuedTuples.@values

ValuedTuples.ValuedTuple — Type.

julia> using ValuedTuples

julia> ValuedTuple((1, 2, 3), @values a b c)
(@VT a = 1 b = 2 c = 3)

julia> ValuedTuple((1, 2, 3), @values a)
ERROR: Tuple size mismatch: leftovers (2, 3)
[...]

julia> ValuedTuple((1,), @values a b c)
ERROR: Tuple size mismatch: leftovers (Val{:b}(), Val{:c}())
[...]

julia> ValuedTuple((1, 2), (3, 4))
ERROR: All names must be Vals
[...]

source

ValuedTuples.delete — Method.

delete(v::ValuedTuple, value)

julia> using ValuedTuples

julia> delete((@VT a = 1 b = 2), @value a)
(@VT b = 2)

source

ValuedTuples.match_index — Method.

match_index(v::ValuedTuple, value)

julia> using ValuedTuples

julia> v = @VT a = 1 b = 2 a = 3
(@VT a = 1 b = 2 a = 3)

julia> match_index(v, @value a)
(1, 3)

source

ValuedTuples.value_names — Method.

value_names(t::Type{T}) where T <: ValuedTuple

Extract the names of a valued tuple exclusively based on its type.

julia> using ValuedTuples

julia> v = @VT a = 1 b = 2;

julia> value_names(typeof(v))
(:a, :b)

source

ValuedTuples.@VT — Macro.

@VT args...

Make a ValuedTuple. A valued tuple can be indexed only with Vals (create with @value). Valued tuples can be manipulated in a type-stable way because the names are directly encoded into the type. You can use repeated values. getindex will take the last match when trying to index at a repeated value; for all matches, use match_index instead. A vector of NamedTuples with consistent names will conveniently print as a markdown table.

julia> using ValuedTuples

julia> b = 2;

julia> v = @VT a = 1 b a = 3
(@VT a = 1 b = 2 a = 3)

julia> v[@value b]
2

julia> v[@value d]
ERROR: BoundsError: attempt to access ()
[...]

julia> v[@value a]
3

julia> @VT x * y
ERROR: Unable to decompose assignment x * y
[...]

julia> [(@VT a = 1 b = 2), (@VT b = 2 c = 3)]
2-element Array{ValuedTuples.ValuedTuple{Tuple{Int64,Int64},N} where N<:Tuple,1}:
 (@VT a = 1 b = 2)
 (@VT b = 2 c = 3)

julia> t = [(@VT a = 1 b = 2), (@VT a = 3 b = 4)]
| a   | b   |
|:--- |:--- |
| 1   | 2   |
| 3   | 4   |

source

ValuedTuples.@value — Macro.

julia> using ValuedTuples

julia> @value a
Val{:a}()

source

ValuedTuples.@values — Macro.

julia> using ValuedTuples

julia> @values a b
(Val{:a}(), Val{:b}())

source

Base.merge — Method.

Base.merge(v1::ValuedTuple, v2::ValuedTuple)

julia> using ValuedTuples

julia> merge((@VT a = 1 b = "b"), (@VT c = 3 d = "d"))
(@VT a = 1 b = "b" c = 3 d = "d")

source